# frontend/public/openapi.json
{
  "components": {
    "schemas": {
...
      "CsvRowError": {
        "properties": {
          "line": {
            "title": "Line",
            "type": "integer"
          },
          "raw": {
            "additionalProperties": true,
            "title": "Raw",
            "type": "object"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "line",
          "reason"
        ],
        "title": "CsvRowError",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuestionBankIn": {
        "description": "可复用题库(design contract):一份题目集合,可被多个 Agent / 会议挂载。\n\n题库 MUST NOT 含 rubric / engine / 人设(那些是 Agent 的职责)。",
        "properties": {
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "QuestionBankIn",
        "type": "object"
      },
      "QuestionBankOut": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by_admin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Admin"
          },
          "created_by_client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Client"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "question_bank_id": {
            "title": "Question Bank Id",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "title": "Questions",
            "type": "array"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "updated_by_admin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By Admin"
          },
          "updated_by_client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By Client"
          },
          "version": {
            "default": "v1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "question_bank_id"
        ],
        "title": "QuestionBankOut",
        "type": "object"
      },
      "QuestionBankUploadResult": {
        "description": "题库 CSV 批量上传结果(spec:免逐题手加)。逐行校验:合法行入库、非法行跳过 + 错误明细。",
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/CsvRowError"
            },
            "title": "Errors",
            "type": "array"
          },
          "imported": {
            "title": "Imported",
            "type": "integer"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "question_bank_id": {
            "title": "Question Bank Id",
            "type": "string"
          },
          "rejected": {
            "title": "Rejected",
            "type": "integer"
          },
          "total_questions": {
            "title": "Total Questions",
            "type": "integer"
          },
          "total_rows": {
            "title": "Total Rows",
            "type": "integer"
          }
        },
        "required": [
          "question_bank_id",
          "mode",
          "total_rows",
          "imported",
          "rejected",
          "total_questions"
        ],
        "title": "QuestionBankUploadResult",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "VivaVoce control plane (real-time voice oral-exam). All /api/* endpoints require authentication.",
    "title": "VivaVoce Orchestrator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/integration/question-banks": {
      "get": {
        "operationId": "api_list_question_banks_api_integration_question_banks_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionBankOut"
                  },
                  "title": "Response Api List Question Banks Api Integration Question Banks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api List Question Banks",
        "tags": [
          "integration-api"
        ]
      },
      "post": {
        "operationId": "api_create_question_bank_api_integration_question_banks_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBankIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Create Question Bank",
        "tags": [
          "integration-api"
        ]
      }
    },
    "/api/integration/question-banks/{question_bank_id}": {
      "delete": {
        "description": "机器删题库(复用 assert_qb_deletable:活动会话 / Agent 默认 / 时段 Slot 引用 → 409)。",
        "operationId": "api_delete_question_bank_api_integration_question_banks__question_bank_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Delete Question Bank",
        "tags": [
          "integration-api"
        ]
      },
      "get": {
        "operationId": "api_get_question_bank_api_integration_question_banks__question_bank_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Get Question Bank",
        "tags": [
          "integration-api"
        ]
      },
      "put": {
        "description": "机器改题库(复用 db.update_question_bank 版本快照)。",
        "operationId": "api_update_question_bank_api_integration_question_banks__question_bank_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBankIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Update Question Bank",
        "tags": [
          "integration-api"
        ]
      }
    },
    "/api/integration/question-banks/{question_bank_id}/upload-csv": {
      "post": {
        "description": "机器批量上传题库 CSV(复用 parse_questions_csv + replace 0 题保护;body = 原始 CSV 文本 UTF-8)。",
        "operationId": "api_upload_csv_api_integration_question_banks__question_bank_id__upload_csv_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "append",
              "pattern": "^(append|replace)$",
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankUploadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Upload Csv",
        "tags": [
          "integration-api"
        ]
      }
    },
...
    "/api/question-banks": {
      "get": {
        "operationId": "list_question_banks_api_question_banks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionBankOut"
                  },
                  "title": "Response List Question Banks Api Question Banks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Question Banks",
        "tags": [
          "question-banks"
        ]
      },
      "post": {
        "operationId": "create_question_bank_api_question_banks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBankIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Question Bank",
        "tags": [
          "question-banks"
        ]
      }
    },
    "/api/question-banks/csv-template": {
      "get": {
        "description": "下载题库 CSV 模板(表头 + 一行示例)。text 必填,reference_answer/weight/difficulty\n可留空用默认(weight=1.0 / difficulty=3)。design contract:题目级 follow_up 已废弃,模板不含该列。",
        "operationId": "download_csv_template_api_question_banks_csv_template_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Csv Template",
        "tags": [
          "question-banks"
        ]
      }
    },
    "/api/question-banks/{question_bank_id}": {
      "delete": {
        "description": "删除题库(design contract,仅 admin)。引用完整性检查(409):\n\n被以下任一引用时挡删 —— ① 活动会话(question_bank_id 字段);② 任何时段池 Slot;\n③ 任何 Agent 的 default_question_bank_id(否则留悬挂默认)。\n已固化 resolved_questions 的(终态/历史)会话不算引用(快照独立)。",
        "operationId": "delete_question_bank_api_question_banks__question_bank_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Question Bank",
        "tags": [
          "question-banks"
        ]
      },
      "get": {
        "operationId": "get_question_bank_api_question_banks__question_bank_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Question Bank",
        "tags": [
          "question-banks"
        ]
      },
      "put": {
        "description": "改版不覆盖历史(design contract):当前版本入 version_history、version bump。",
        "operationId": "update_question_bank_api_question_banks__question_bank_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBankIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Question Bank",
        "tags": [
          "question-banks"
        ]
      }
    },
    "/api/question-banks/{question_bank_id}/upload-csv": {
      "post": {
        "description": "批量上传题库 CSV(免逐题手加)。body = 原始 CSV 文本(UTF-8)。逐行校验:合法行入库、\n非法行跳过并返回错误明细。mode=append 追加到现有题目;mode=replace 整批替换题库题目。\n改版语义沿用 update_question_bank(当前版本入 version_history、version bump)。",
        "operationId": "upload_csv_api_question_banks__question_bank_id__upload_csv_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "append",
              "pattern": "^(append|replace)$",
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankUploadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Csv",
        "tags": [
          "question-banks"
        ]
      }
    },
    "/api/question-banks/{question_bank_id}/versions": {
      "get": {
        "operationId": "get_question_bank_versions_api_question_banks__question_bank_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "question_bank_id",
            "required": true,
            "schema": {
              "title": "Question Bank Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionBankVersionsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Question Bank Versions",
        "tags": [
          "question-banks"
        ]
      }
    },
...
  }
}